Title: 우리 지금 만나 당장 만나, Time: 1110ms, Memory: 53MB, Status: Wrong Answer - Codetree

+++ 250618/우리 지금 만나 당장 만나/meet-now-right-now.java
@@ -49,7 +49,7 @@ public class Main {
             double max = -1;
             int index = -1;
             for (int i = 0; i<n; i++) {
-                double x = (double) ((Math.abs(person[i].start - mid)) / person[i].speed);
+                double x = ((Math.abs(person[i].start - mid)) / person[i].speed);
                 if (x > max) {
                     max = x;
                     index = i;
